+++ examples/llms/providers/customProvider.ts
@@ -86,7 +86,7 @@ export class CustomLLM extends LLM<CustomLLMOutput, CustomGenerateOptions> {
 
   protected async _generate(
     input: LLMInput,
-    options: CustomGenerateOptions,
+    options: Partial<CustomGenerateOptions>,
     run: GetRunContext<this>,
   ): Promise<CustomLLMOutput> {
     // this method should do non-stream request to the API
@@ -101,7 +101,7 @@ export class CustomLLM extends LLM<CustomLLMOutput, CustomGenerateOptions> {
 
   protected async *_stream(
     input: LLMInput,
-    options: CustomGenerateOptions,
+    options: Partial<CustomGenerateOptions>,
     run: GetRunContext<this>,
   ): AsyncStream<CustomLLMOutput, void> {
     // this method should do stream request to the API