+++ src/adapters/vertexai/llm.ts
@@ -23,6 +23,7 @@ import {
   GenerateOptions,
   LLMCache,
   LLMMeta,
+  StreamGenerateOptions,
 } from "@/llms/base.js";
 import { shallowCopy } from "@/serializer/utils.js";
 import type { GetRunContext } from "@/context.js";
@@ -134,7 +135,7 @@ export class VertexAILLM extends LLM<VertexAILLMOutput> {
 
   protected async *_stream(
     input: LLMInput,
-    options: GenerateOptions | undefined,
+    options: Partial<StreamGenerateOptions>,
     run: GetRunContext<this>,
   ): AsyncStream<VertexAILLMOutput, void> {
     const generativeModel = createModel(