+++ src/tools/database/elasticsearch.ts
@@ -22,6 +22,7 @@ import {
   BaseToolRunOptions,
   JSONToolOutput,
   ToolInputValidationError,
+  CustomToolEmitter,
 } from "@/tools/base.js";
 import { Cache } from "@/cache/decoratorCache.js";
 import { RunContext } from "@/context.js";
@@ -30,6 +31,7 @@ import { ValidationError } from "ajv";
 import { AnyToolSchemaLike } from "@/internals/helpers/schema.js";
 import { parseBrokenJson } from "@/internals/helpers/schema.js";
 import { Client, ClientOptions, estypes as ESTypes } from "@elastic/elasticsearch";
+import { Emitter } from "@/emitter/emitter.js";
 
 export interface ElasticSearchToolOptions extends BaseToolOptions {
   connection: ClientOptions;
@@ -96,6 +98,14 @@ export class ElasticSearchTool extends Tool<
     });
   }
 
+  public readonly emitter: CustomToolEmitter<
+    ToolInput<this>,
+    JSONToolOutput<ElasticSearchToolResult>
+  > = Emitter.root.child({
+    namespace: ["tool", "database", "elasticsearch"],
+    creator: this,
+  });
+
   protected validateInput(
     schema: AnyToolSchemaLike,
     input: unknown,
@@ -145,7 +155,7 @@ export class ElasticSearchTool extends Tool<
 
   protected async _run(
     input: ToolInput<this>,
-    _options: BaseToolRunOptions | undefined,
+    _options: Partial<BaseToolRunOptions>,
     run: RunContext<this>,
   ): Promise<JSONToolOutput<any>> {
     if (input.action === ElasticSearchAction.ListIndices) {